Enable multi-objective optimisation
Multiple scalars can now be returned by an objective function,
while the default is only one scalar.

In case of multiple objectives (multiple numbers), the
objective function must be decorated with the respective directions:

@directions(['minimize', 'maximize'])
def my_objective(...):
  # do your calculations
  return value_to_minimize, value_to_maximize

If only one number is returned, the decorator can be omitted if that number
should be minimized.

If maximization is desired,n single-objective optimisation, the decorator
should be used with
@directions(['maximize'])
def my_objective(...):
  # do your calculations
  return value_to_maximize

All plotting functions of the Inspector now have a keyword argument
objective_number
to indicate for which of those objective the respctive plot should be created.